# packages/contracts/openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: SuperAssistant API
  version: 1.1.0
  description: MVP v1 baseline APIs with core admin endpoints.
servers:
  - url: /api/v1
components:
  parameters:
    XRequestId:
      in: header
      name: X-Request-Id
      required: true
      schema:
        type: string
    XAppVersion:
      in: header
      name: X-App-Version
      required: true
      schema:
        type: string
    XPlatform:
      in: header
      name: X-Platform
      required: true
      schema:
        type: string
        enum: [ios, android, web-admin]
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /suggestions/generate:
    post:
      summary: Generate purchase suggestions
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [AUTO, MANUAL]
      responses:
        '200': { description: success }
  /suggestions/latest:
    get:
      summary: Get latest suggestion
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
      responses:
        '200': { description: success }
  /push/device/register:
    post:
      summary: Register push device
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, device_token]
              properties:
                platform: { type: string }
                device_token: { type: string }
      responses:
        '200': { description: success }
  /push/preference:
    put:
      summary: Update push preference
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled, notify_time]
              properties:
                enabled: { type: boolean }
                notify_time: { type: string }
      responses:
        '200': { description: success }
...